requirements.txt
mysql-connector-python~=8.0.33
tqdm~=4.65.0